+++ packages/account-node/src/index.ts
@@ -1,5 +1,6 @@
 import AutoAccount from "./auto"
 import LimitedAccount from "./limited"
+import ContextAccount from "./contextAccount"
 import AuthContextAccount from "./authcontext"
 import VeridaDatabaseAuthType from "./authTypes/VeridaDatabase"
 export * from './nodeSelector'
@@ -8,5 +9,6 @@ export {
     AutoAccount,
     VeridaDatabaseAuthType,
     LimitedAccount,
+    ContextAccount,
     AuthContextAccount
 }