+++ packages/cli-tools/src/commands/interfaces.ts
@@ -1,0 +1,15 @@
+
+export interface SendInboxMessageOptions {
+    did: string
+    message: string
+    network: string
+    privateKey: string
+    subject: string
+
+    sendContext: string
+    receiveContext: string
+}
+
+export interface GetDIDDocumentOptions {
+    did: string
+}